Big changes in MenuRepository and PersonRepository.
I fixed some problems, but now i added a new things.
In MenuRepository, I fixed if i could to say for some reason, but well... i brought a new print, this print give you what you need to put in it, i fixed a problem for the 'ic', this was at two sites in the same time. All of these change on a swtich.

In PersonRepository, i fixed some errors and mistakes that i put before, like on the method 'getValidateAge' i made a variable but this variable was temporary, for that when you call this variable you never get a variable with something on it, upside down, you have it but empty or void/null. So for that i called the 'Person' 'age' and 'genre' properties to get something in the variable. Same thing in 'getValidateGenre'. I added a new method, 'EditPerson', this serves like when you did something wrong putting your dates in the console, you can leave it as it was or edit it.