Important update in OptionRepository and little update in App.
In OptionRepository, i fixed a problem with the 'while' loop, this mistake causes a infinite loop in the 'while', for that you couldn't to put an input in the console; you couldn't to stop it, so for that big mistake i fixed all methods. Other things to pay attetion is the 'while' in the 'while' on it, this had a 'scanner.nextline()' and this causes a ghost space, other thing is when this happening you lose your sight on it, so for that now when this happen you calling the menu, it's easier like this.