Big update in OptionRepository.
I fixed a problem with 'sign in option' it had an error; when you try to put a number different to the number 1 appears a ghost space, but it doesn't happening when you putted a letter. I added a new switches (methods) and i found out the switch menus need a 'while'.