I'm losing my mind...
I fixed some errors but i was trying to fix one special; it doesn't work when you input a password to delete someone you want to delete by password.
I added new methods but it doesn't matter. There's another problem when the console get a person, if this get another remove the before and it get the new.