Little changes in Menu, Option, PersonRepository and App.
In menu i just added more text in the print.

In OptionRe i moved the introduction method up.

In PersonRe i fixed a problem with the incialization object, this get changes but never add another object, perhaps the object was called everytime, when this should be different.

In App i don't know what i did. I was just testing.