Update in style.TS, new icon and image, and edit in index.HTML.
In Style, i added a style on the page, you can watch the color, the change color, the buttons, and so on.

New icon for the page tab and image for the banner.

In index.HTML, i edit some things for the style, if i never change, some style wouldn't be there.